+++ src/pages-layout/HeaderWrapper/index.ts
@@ -1,0 +1,1 @@
+export { HeaderWrapper } from './HeaderWrapper'
